Define state default in the reducers
